Modernize .NET tooling and enable AI workflows
- See SDK_MIGRATION.md for details
- **64-bit only** architecture enforcement (x86/Win32 removed)
- **Registration-free COM** implementation (Native + Managed)
- **Unified launcher**: FieldWorks.exe replaced the historical LexText.exe stub across build, installer, and documentation
- **MSBuild Traversal SDK** for declarative builds
- **Test framework modernization** (RhinoMocks → Moq, NUnit 4 ready)
- **Central Package Management (CPM)** via `Directory.Packages.props`
- **Unified test runner** (`test.ps1`) for managed and native tests
- **Stale DLL detection** via single-pass pre-build validation
- **Installer validation tooling** with snapshot-based evidence collection
- **Binding redirect cleanup** — eliminated manual `<bindingRedirect>` entries
- **Developer environment tooling** (Defender exclusions, dependency verification)
- **AGENTS.md documentation convention** for AI agent and developer onboarding

+++ Lib/src/graphite2/src/Collider.cpp
@@ -29,6 +29,7 @@ of the License or (at your option) any later version.
 #include <cmath>
 #include <string>
 #include <functional>
+#include <cstdint>
 #include "inc/Collider.h"
 #include "inc/Segment.h"
 #include "inc/Slot.h"
@@ -519,7 +520,7 @@ bool ShiftCollider::mergeSlot(Segment *seg, Slot *slot, const SlotCollision *csl
 
 #if !defined GRAPHITE2_NTRACING
                     if (dbgout)
-                        dbgout->setenv(1, reinterpret_cast<void *>(j));
+                        dbgout->setenv(1, reinterpret_cast<void *>(static_cast<std::intptr_t>(j)));
 #endif
                     if (omin > otmax)
                         _ranges[i].weightedAxis(i, vmin - lmargin, vmax + lmargin, 0, 0, 0, 0, 0,
@@ -538,7 +539,7 @@ bool ShiftCollider::mergeSlot(Segment *seg, Slot *slot, const SlotCollision *csl
             {
 #if !defined GRAPHITE2_NTRACING
                     if (dbgout)
-                        dbgout->setenv(1, reinterpret_cast<void *>(-1));
+                        dbgout->setenv(1, reinterpret_cast<void *>(static_cast<std::intptr_t>(-1)));
 #endif
                 isCol = true;
                 if (omin > otmax)
